refactor: Wave 2 - extract server-side spec construction hooks to Core/Net/
Split RegisterSpecConstructionHooks() (cost scaling, grid/Extend expansion +
group bookkeeping, placement leniency, pre-BeginPlay group membership, and the
chain-integrity guards) and its file-local GSFActiveSpecGroupProxy static out of
SFGameInstanceModule.cpp into Core/Net/SFGameInstanceModule_SpecHooks.cpp.

Verbatim partial-class split: same USFGameInstanceModule member, registered
identically from StartupModule; the static is self-contained to this function.
SFGameInstanceModule.cpp drops 978->329 lines (now just lifecycle dispatch +
blueprint/belt-support construct hooks). Shipping compile clean (Result: Succeeded).
